+++ lib/mini_kraken/core/disj2.rb
@@ -8,13 +8,13 @@ require_relative 'outcome'
 
 module MiniKraken
   module Core
-    # The conjunction is a relation that accepts only goal(s) as its two
-    # arguments. It succeeds if and only both its goal arguments succeeds.
-    class Conj2 < GoalRelation
+    # The disjunction is a relation that accepts only goal(s) as its two
+    # arguments. It succeeds if at least one of its goal arguments succeeds.
+    class Disj2 < GoalRelation
       include Singleton
 
       def initialize
-        super('conj2', nil)
+        super('disj2', nil)
       end
 
       # @param actuals [Array<Term>] A two-elements array
@@ -22,18 +22,18 @@ module MiniKraken
       # @return [Fiber<Outcome>] A Fiber that yields Outcomes objects
       def solver_for(actuals, anEnv)
         g1, g2 = *validated_args(actuals)
-        Fiber.new { conjunction(g1, g2, anEnv) }
+        Fiber.new { disjunction(g1, g2, anEnv) }
       end
 
-      # Yields [Outcome, NilClass] result of the conjunction
+      # Yields [Outcome, NilClass] result of the disjunction
       # @param g1 [Goal] First goal argument
       # @param g2 [Goal] Second goal argument
       # @param voc [Vocabulary] A vocabulary object
-      def conjunction(g1, g2, voc)
+      def disjunction(g1, g2, voc)
         # require 'debug'
         outcome1 = nil
         outcome2 = nil
-        if g1.relation.kind_of?(Fail) || g2.relation.kind_of?(Fail)
+        if g1.relation.kind_of?(Fail) && g2.relation.kind_of?(Fail)
           Fiber.yield Outcome.new(:"#u", voc)
         else
           f1 = g1.attain(voc)
@@ -43,26 +43,20 @@ module MiniKraken
 
             outcome1.parent = voc unless outcome1.parent
             if outcome1.successful?
-              f2 = g2.attain(outcome1)
-              loop do
-                outcome2 = f2.resume
-                break unless outcome2
-
-                outcome2.parent = voc unless outcome2.parent
-                if outcome2.successful?
-                  res = Outcome.new(:"#s", voc)
-                  res.merge(outcome1)
-                  res.merge(outcome2)
-                  Fiber.yield res
-                else
-                  Fiber.yield outcome2
-                end
-                outcome2.clear
-              end
-            else
               Fiber.yield outcome1
+              outcome1.clear
             end
-            voc.clear if outcome1&.successful? && outcome2&.successful?
+          end
+          f2 = g2.attain(voc)
+          loop do
+            outcome2 = f2.resume
+            break unless outcome2
+
+            outcome2.parent = voc unless outcome2.parent
+            if outcome2.successful?
+              Fiber.yield outcome2
+              outcome2.clear
+            end
           end
         end